Assignment 6
Question 1 :
A bank has a token system where customers take a token and are served in the order they arrived. You need to:
Add new customers to the queue.
Serve customers in a First In, First Out (FIFO) order.
Question 2 :
Check who is next in line without removing them.
A university is conducting an event where students register for workshops. However, a student can only register once for a particular workshop.
You need to store unique student IDs for each workshop.
If a student tries to register again, the system should prevent duplicates. Implement with C#